Added dungeon progression
Added the ability to move to different floors of the dungeon, which involved reworking several map methods. Additionally, I think I added a few new potions, although it's possible that I added those in a previous update and forgot. Those potions would be the Antidote, Permafrost Potion, and Firebrew Potion.