Improved UI, AI, and added ranged comat
Added the entire framework for ranged combat. The player currently spawns with a light crossbow and two types of bolts. I replaced the placeholder AI with a more versatile AI. The UI now shows the secondary weapon, if one is equipped. Monsters who have a ranged weapon equipped should be able to fire them.